move api file into helpers, baseURL to use production
